Perform derivation check in the parent of the tab-related components
