+++ src/components/shared/Entity/Details/Overview/index.tsx
@@ -1,6 +1,5 @@
 import { Grid } from '@mui/material';
 import { DataPreview } from 'components/collection/DataPreview';
-import { useEditorStore_currentCatalog } from 'components/editor/Store/hooks';
 import { TaskEndpoints } from 'components/shared/TaskEndpoints';
 import { useEntityType } from 'context/EntityContext';
 import useGlobalSearchParams, {
@@ -8,7 +7,7 @@ import useGlobalSearchParams, {
 } from 'hooks/searchParams/useGlobalSearchParams';
 import { useLiveSpecs_details } from 'hooks/useLiveSpecs';
 import { useMemo } from 'react';
-import { hasLength, specContainsDerivation } from 'utils/misc-utils';
+import { hasLength } from 'utils/misc-utils';
 import ShardInformation from '../../Shard/Information';
 import Usage from '../Usage';
 import DetailsSection from './DetailsSection';
@@ -16,22 +15,17 @@ import DetailsSection from './DetailsSection';
 // TODO (details page)
 // Temporary - allow to pass in the name
 interface Props {
+    isDerivation: boolean;
     name?: string;
 }
 
-function Overview({ name }: Props) {
+function Overview({ isDerivation, name }: Props) {
     const entityType = useEntityType();
     const isCollection = entityType === 'collection';
     const catalogName = useGlobalSearchParams(GlobalSearchParams.CATALOG_NAME);
     const entityName = name ?? catalogName;
     const { liveSpecs, isValidating: validatingLiveSpecs } =
         useLiveSpecs_details(entityType, entityName);
-
-    const currentCatalog = useEditorStore_currentCatalog({
-        localScope: true,
-    });
-    const catalogSpec = currentCatalog?.spec ?? null;
-    const { isDerivation } = specContainsDerivation(catalogSpec);
 
     const latestLiveSpec = useMemo(
         () =>
